requirements.txt
flask
# vanna[snowflake]
db-dtypes
python-dotenv
vanna[chromadb,openai]